+++ src/components/PhotoList.tsx
@@ -1,18 +1,11 @@
 "use client";
 
 import React from "react";
-import NextImage from "next/image"; // next/imageを別名でインポート
 import { useEffect, useState } from "react";
 import EXIF from "exif-js";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Button } from "@/components/ui/button";
-import {
-  IconApeture,
-  IconCamera,
-  IconExposure,
-  IconIso,
-  IconLength,
-} from "./Icon";
+import PhotoDetail from "./PhotoDetail";
 
 type Photo = {
   id: string;
@@ -36,7 +29,6 @@ type ExifData = {
 
 const PhotoList = () => {
   const [photoList, setPhotoList] = useState<Photo[]>([]);
-  const [exifData, setExifData] = useState<Record<string, ExifData>>({});
   const [selectedGenres, setSelectedGenres] = useState<string[]>([]); // 選択されたジャンル
   const [filteredPhotos, setFilteredPhotos] = useState<Photo[]>([]); // 絞り込んだ写真リスト
 
@@ -58,39 +50,6 @@ const PhotoList = () => {
     });
     setFilteredPhotos(updatedPhotos);
   }, [selectedGenres, photoList]);
-
-  const fetchExifData = (photo: Photo) => {
-    const img = new window.Image(); // ネイティブのImageオブジェクトを明示的に使用
-    img.crossOrigin = "Anonymous";
-    img.src = photo.image.url;
-
-    img.onload = () => {
-      EXIF.getData(img, function (this: any) {
-        const cameraMake = String(EXIF.getTag(this, "Make") || "N/A");
-        const cameraModel = String(EXIF.getTag(this, "Model") || "N/A");
-        const exposureTime = String(EXIF.getTag(this, "ExposureTime") || "N/A");
-        const focalLength = String(EXIF.getTag(this, "FocalLength") || "N/A");
-        const aperture = String(EXIF.getTag(this, "FNumber") || "N/A");
-        const iso = String(EXIF.getTag(this, "ISOSpeedRatings") || "N/A");
-
-        setExifData((prev) => ({
-          ...prev,
-          [photo.id]: {
-            cameraMake,
-            cameraModel,
-            exposureTime,
-            focalLength,
-            aperture,
-            iso,
-          },
-        }));
-      });
-    };
-
-    img.onerror = () => {
-      console.error(`画像の読み込みに失敗しました: ${photo.image.url}`);
-    };
-  };
 
   // ジャンルごとのユニークなキーを生成
   const genreWithIds = photoList.map((photo) => ({
@@ -153,65 +112,7 @@ const PhotoList = () => {
 
       <div className='px-2 columns-1 gap-0 space-y-3 md:columns-3 md:gap-3 md:space-y-3 md:px-2 lg:columns-4 lg:gap-4 lg:space-y-4 lg:px-4'>
         {filteredPhotos.map((photo) => (
-          <dl
-            key={photo.id}
-            className='border border-gray-200 shadow-md rounded-md break-inside-avoid'
-          >
-            <dt>
-              <NextImage
-                className='w-full rounded-t-md'
-                src={photo.image.url}
-                alt='description'
-                width={photo.image.width}
-                height={photo.image.height}
-                onLoadingComplete={() => fetchExifData(photo)} // Exif情報取得
-              />
-            </dt>
-            <dd className='p-4 space-y-1'>
-              <dl className='flex items-center'>
-                <dt>
-                  <IconCamera color='black' label='Camera Maker' />
-                </dt>
-                <dd className='ml-2'>
-                  {`${exifData[photo.id]?.cameraMake || "Loading..."} / ${
-                    exifData[photo.id]?.cameraModel || "Loading..."
-                  }`}
-                </dd>
-              </dl>
-              <dl className='flex items-center'>
-                <dt>
-                  <IconExposure color='black' label='Exposure' />
-                </dt>
-                <dd className='ml-2'>
-                  {`${exifData[photo.id]?.exposureTime || "Loading..."} sec`}
-                </dd>
-              </dl>
-              <dl className='flex items-center'>
-                <dt>
-                  <IconLength color='black' label='Focal Length' />
-                </dt>
-                <dd className='ml-2'>
-                  {`${exifData[photo.id]?.focalLength || "Loading..."} mm`}
-                </dd>
-              </dl>
-              <dl className='flex items-center'>
-                <dt>
-                  <IconApeture color='black' label='Apature' />
-                </dt>
-                <dd className='ml-2'>
-                  {`f / ${exifData[photo.id]?.aperture || "Loading..."}`}
-                </dd>
-              </dl>
-              <dl className='flex items-center'>
-                <dt>
-                  <IconIso color='black' label='iso' size={24} />
-                </dt>
-                <dd className='ml-2'>
-                  {exifData[photo.id]?.iso || "Loading..."}
-                </dd>
-              </dl>
-            </dd>
-          </dl>
+          <PhotoDetail key={photo.id} photo={photo} />
         ))}
       </div>
     </>
